Couldn't figure out waiting for fetch before continuing function, so changed code to start empty, no default weather data, and display and save search history in the fetch forecast function directly. This ensures that cities will only go into search history if they return valid results.
Added functionality to search history buttons.

added loop to check if search already exists in history, and if so splices the existing entry and adds it back to the end of the array.

Reversed order in which searchHistory loads, as it makes more sense to me for recent searches to be at the top. This probably could have just been done in HTML though...